Added properties
