+++ src/test/resources/config.properties
@@ -1,0 +1,5 @@
+url=https://www.saucedemo.com/
+browser =chrome
+username=standard_user
+password=secret_sauce
+timeout=10
